move match candidaates to helper function, add source to newMessage, limit 1 on company query
