+++ server/db/queries/companies/getRecommendedPositions.js
@@ -2,42 +2,41 @@ const { singleQuery } = require('../../baseQueries');
 
 const getRecommendedPositions = async () => {
   const queryText = `
-  WITH all_positions as (
-    SELECT
-      c.company_name,
-      c.logo,
-      cp.company_id,
-      cp.created_date,
-      cp.id,
-      l.formatted_address,
-      q.question_key,
-      JSONB_BUILD_OBJECT(
-        'name', t.name,
-        'level', pts.level
-        )
-      AS user_languages,
-      CASE WHEN q.question_key = 'role' THEN qr.value END AS roles,
-      CASE WHEN q.question_key != 'role' THEN max(coalesce(uqr.value, qr.value)) END AS question_responses
-    FROM company_positions cp
-    JOIN companies c ON c.id = cp.company_id
-    JOIN user_question_responses uqr ON uqr.position_id = cp.id
-    JOIN locations l on l.position_id = cp.id
-    JOIN question_responses qr ON qr.id = uqr.response_id
-    JOIN questions q ON q.id = uqr.question_id
-    JOIN position_tech_stack pts ON pts.position_id = cp.id
-    JOIN technologies t ON pts.technology_id = t.id
-    WHERE q.question_key in ('post_to_job_board', 'role', 'salary', 'title')
-    GROUP BY
-      c.company_name,
-      c.logo,
-      cp.company_id,
-      cp.created_date,
-      cp.id,
-      l.formatted_address,
-      pts.level,
-      q.question_key,
-      qr.value,
-      t.name
+    WITH all_positions as (
+      SELECT
+        c.company_name,
+        c.logo,
+        cp.company_id,
+        cp.id,
+        l.formatted_address,
+        q.question_key,
+        MAX(uqr.created_date) AS created_date,
+        JSONB_BUILD_OBJECT(
+          'name', t.name,
+          'level', pts.level
+          )
+        AS user_languages,
+        CASE WHEN q.question_key = 'role' THEN qr.value END AS roles,
+        CASE WHEN q.question_key != 'role' THEN max(coalesce(uqr.value, qr.value)) END AS question_responses
+      FROM company_positions cp
+      JOIN companies c ON c.id = cp.company_id
+      JOIN user_question_responses uqr ON uqr.position_id = cp.id
+      JOIN locations l on l.position_id = cp.id
+      JOIN question_responses qr ON qr.id = uqr.response_id
+      JOIN questions q ON q.id = uqr.question_id
+      JOIN position_tech_stack pts ON pts.position_id = cp.id
+      JOIN technologies t ON pts.technology_id = t.id
+      WHERE q.question_key in ('post_to_job_board', 'role', 'salary', 'title')
+      GROUP BY
+        c.company_name,
+        c.logo,
+        cp.company_id,
+        cp.id,
+        l.formatted_address,
+        pts.level,
+        q.question_key,
+        qr.value,
+        t.name
     )
     SELECT
       ap.company_id  AS "companyId",
