+++ server/graphql/resolvers/helpers/index.js
@@ -1,0 +1,5 @@
+const matchCandidate = require('./matchCandidate');
+
+module.exports = {
+  matchCandidate,
+};
